requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
psycopg2-binary==2.9.3
mysql-connector-python==8.0.26
jieba==0.42.1
python-dotenv==0.19.0 